Add toJson/fromJson and toDDL/fromDDL to PType
- Add Jackson dependency to partiql-cli
- Add toJson()/fromJson() instance/static methods on PType for JSON serialization
- Add toDDL()/fromDDL() instance/static methods on PType for DDL serialization
- Support all 28 PType variants including INTERVAL_YM and INTERVAL_DT
