+++ partiql-spi/build.gradle.kts
@@ -21,7 +21,6 @@ plugins {
 
 dependencies {
     api(Deps.ionElement)
-    implementation(Deps.jacksonDatabind)
     implementation(Deps.kotlinxCollections)
     testImplementation(Deps.kasechange)
     testImplementation(Deps.snapshotTest)
